Add option ENABLE_VORBIS_PSY
Enable Vorbis psychoacoustic model for the Speex codec.
This is an experimental feature in the Speex codec that improves sound quality.
It allows the Speex codec to use the Vorbis psychoacoustic model for better speech compression,
especially at higher bitrates.
This model helps identify which parts of the sound are less noticeable to the human ear
and compresses them more, while preserving overall quality.